requirements.txt
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.3
numpy==1.26.4
pandas==2.2.1
dash==2.14.2
dash-bootstrap-components==1.5.0
msal==1.27.0
tqdm==4.66.2